refactor: mbit 폴더 수정 game, loading, result로 구분
