+++ src/app/_mbti/_game/bubble.tsx
