+++ src/app/_mbti/_game/mbti-data.ts
