+++ src/app/_mbti/_game/mbti-game-screen.tsx
